refactor: take Device out of VulkanEngine
